requirements.txt
# Data Analysis Packages
pandas>=2.0.0
numpy>=1.24.0

# Data Profiling
ydata-profiling>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.24.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.22.0

# Data Processing
scipy>=1.10.0
scikit-learn>=1.2.0

# Environment Variables
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0